requirements.txt
Django==1.9.7
dj_database_url==0.4.2
djangorestframework==3.2.3
numpy==1.12.0
scikit-learn==0.18.1